refactor(workflow): extract shared writeWorkflowVersionAndMirror helper
The transaction + mirror + post-commit invalidate boilerplate is now a
single reusable method on WorkflowVersionCoreSyncService; each funnel
supplies only its workspace write via a callback. updateWorkflowVersion
StepsAndTrigger becomes a one-liner over it. This is the foundation for
wiring the remaining upsert funnels (edges, draft/duplicate/positions,
create hooks, AI tool, activation) before removing the listener's
upsert handlers.
